TASK: Skip symlink tests on windows when symlinking fails
On windows it is not unlikely that symlinking does not work due to missing access rights.
In that case a lot of the Files tests fail without any helpful notice and eventually lead
to the developer ignoring the failures, possibly overlooking real failures.

This change makes such symlink failures on Windows skip the test instead.